% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ geo.match }
\alias{ geo.match }
\alias{ geo.match }
\title{ geo.match }
\description{
A package designed to provide matched data for propensity-score based causal analyses using spatially-explicit data
}
